Move import ordering to prettier and further refactors
