+++ config/jest/jest.config.js
@@ -5,13 +5,13 @@ module.exports = {
   silent: true,
   testRegex: "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
   transform: {
-    "^.+\\.tsx?$": "ts-jest"
+    "^.+\\.tsx?$": [
+      "ts-jest",
+      {
+        tsconfig: join(__dirname, "..", "typescript", "tsconfig.json")
+      }
+    ]
   },
   moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  globals: {
-    "ts-jest": {
-      tsconfig: join(__dirname, "..", "typescript", "tsconfig.json")
-    }
-  },
   testEnvironment: join(__dirname, "jest.environment.js")
 };
