+++ config/prettier/prettier.config.js
@@ -1,11 +1,17 @@
 module.exports = {
   arrowParens: "avoid",
+  bracketSameLine: false,
   bracketSpacing: true,
   embeddedLanguageFormatting: "auto",
   htmlWhitespaceSensitivity: "css",
+  importOrder: ["<THIRD_PARTY_MODULES>", "^[./]"],
+  importOrderCaseInsensitive: true,
+  importOrderGroupNamespaceSpecifiers: true,
+  importOrderSeparation: true,
+  importOrderSortSpecifiers: true,
   insertPragma: false,
-  bracketSameLine: false,
   jsxSingleQuote: false,
+  plugins: ["@trivago/prettier-plugin-sort-imports"],
   printWidth: 80,
   proseWrap: "always",
   quoteProps: "as-needed",
