+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-p5-wrapper",
   "version": "3.3.0",
   "description": "A wrapper component that allows you to utilise P5 sketches within React apps.",
-  "homepage": "https://github.com/jamesrweb/react-p5-wrapper",
+  "homepage": "https://github.com/P5-wrapper/react",
   "license": "MIT",
   "main": "dist/components/index.js",
   "module": "dist/components/index.esm.js",
@@ -56,10 +56,10 @@
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/jamesrweb/react-p5-wrapper.git"
+    "url": "https://github.com/P5-wrapper/react.git"
   },
   "bugs": {
-    "url": "https://github.com/jamesrweb/react-p5-wrapper/issues"
+    "url": "https://github.com/P5-wrapper/react/issues"
   },
   "dependencies": {
     "microdiff": "^1.3.1",
@@ -70,43 +70,43 @@
     "react-dom": ">= 17.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.19.0",
-    "@babel/preset-env": "^7.19.0",
+    "@babel/core": "^7.19.3",
+    "@babel/preset-env": "^7.19.3",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.18.6",
     "@testing-library/react": "^13.4.0",
-    "@types/jest": "^28.1.8",
-    "@types/p5": "^1.4.2",
-    "@types/react": "^18.0.19",
+    "@trivago/prettier-plugin-sort-imports": "^3.3.0",
+    "@types/jest": "^29.1.1",
+    "@types/p5": "^1.4.3",
+    "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.36.2",
-    "@typescript-eslint/parser": "^5.36.2",
+    "@typescript-eslint/eslint-plugin": "^5.39.0",
+    "@typescript-eslint/parser": "^5.39.0",
     "babel-loader": "^8.2.5",
     "canvas": "^2.10.1",
     "css-loader": "^6.7.1",
-    "eslint": "^8.23.1",
+    "eslint": "^8.24.0",
     "eslint-plugin-react": "^7.31.8",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
     "gh-pages": "^4.0.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^28.1.3",
-    "jest-environment-jsdom": "^28.1.3",
-    "jest-environment-jsdom-global": "^3.1.2",
+    "jest": "^29.1.2",
+    "jest-environment-jsdom": "^29.1.2",
+    "jest-environment-jsdom-global": "^4.0.0",
     "prettier": "^2.7.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rimraf": "^3.0.2",
-    "rollup": "^2.79.0",
+    "rollup": "^2.79.1",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.32.1",
+    "rollup-plugin-typescript2": "^0.34.1",
     "style-loader": "^3.3.1",
-    "ts-jest": "^28.0.8",
-    "ts-loader": "^9.3.1",
+    "ts-jest": "^29.0.3",
+    "ts-loader": "^9.4.1",
     "tslib": "^2.4.0",
-    "typescript": "^4.8.3",
+    "typescript": "^4.8.4",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.11.0"
+    "webpack-dev-server": "^4.11.1"
   }
 }
