+++ src/contracts/InputProps.ts
@@ -1,5 +1,5 @@
-import type { Sketch } from "./Sketch";
-import type { SketchProps } from "./SketchProps";
+import { type Sketch } from "./Sketch";
+import { type SketchProps } from "./SketchProps";
 
 export type InputProps<Props extends SketchProps = SketchProps> = Props & {
   sketch?: Sketch<Props>;
