+++ src/contracts/Sketch.ts
@@ -1,5 +1,5 @@
-import type { P5CanvasInstance } from "./P5CanvasInstance";
-import type { SketchProps } from "./SketchProps";
+import { type P5CanvasInstance } from "./P5CanvasInstance";
+import { type SketchProps } from "./SketchProps";
 
 export type Sketch<Props extends SketchProps = SketchProps> = (
   instance: P5CanvasInstance<Props>
