+++ src/contracts/WithChildren.ts
@@ -1,3 +1,3 @@
-import type { ReactNode } from "react";
+import { type ReactNode } from "react";
 
 export type WithChildren<T = unknown> = T & { children?: ReactNode };
