+++ src/utils/propsAreEqual.ts
@@ -1,7 +1,7 @@
 import diff from "microdiff";
 
-import type { P5WrapperProps } from "../contracts/P5WrapperProps";
-import type { SketchProps } from "../contracts/SketchProps";
+import { type P5WrapperProps } from "../contracts/P5WrapperProps";
+import { type SketchProps } from "../contracts/SketchProps";
 
 export function propsAreEqual<Props extends SketchProps = SketchProps>(
   previous: P5WrapperProps<Props>,
