+++ src/utils/removeCanvasInstance.ts
@@ -1,7 +1,7 @@
-import type { MutableRefObject } from "react";
+import { type MutableRefObject } from "react";
 
-import type { P5CanvasInstance } from "../contracts/P5CanvasInstance";
-import type { SketchProps } from "../contracts/SketchProps";
+import { type P5CanvasInstance } from "../contracts/P5CanvasInstance";
+import { type SketchProps } from "../contracts/SketchProps";
 
 export function removeCanvasInstance<Props extends SketchProps = SketchProps>(
   canvasInstanceRef: MutableRefObject<P5CanvasInstance<Props> | null>
